requirements.txt
# ETF链接基金T+0马丁格尔量化交易系统依赖

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据源
# 东方财富网 - 实时和历史行情
requests>=2.31.0
beautifulsoup4>=4.12.0  # HTML解析库

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0

# 配置与日志
pyyaml>=6.0
loguru>=0.7.0

# 数据校验
pydantic>=2.0.0

# HTTP 请求
aiohttp>=3.9.0

# 定时任务
apscheduler>=3.10.0
